Refactored uow and repos/Deleted extra repo contracts
